requirements.txt
urllib3~=2.0.2
requests~=2.30.0
python-dotenv~=1.0.0
slack-bolt~=1.18.0
openai~=0.27.6
PyYAML~=6.0
PyGithub~=1.58.2
pydub~=0.25.1
numba~=0.57.0
hupper~=1.12
openai-whisper~=20230314
pandas~=2.0.1
numpy~=1.24.3
joblib~=1.2.0

scikit-learn~=1.2.2
torch~=2.0.1
transformers~=4.29.2